+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Fix exception when roundabout exit is not correctly found (Issue #89)
+- Option to specify maximum locations for matrix request when using non-standard weightings (Issue #94)
 
 ## [4.4.2] - 2018-01-31
 ### Added